fix(FOROME-953): renamed, regouped stores for ws page
