+++ src/components/popper-table-modal.tsx
@@ -7,7 +7,7 @@ import { observer } from 'mobx-react-lite'
 import { ViewTypeEnum } from '@core/enum/view-type-enum'
 import { useOutsideClick } from '@core/hooks/use-outside-click'
 import { t } from '@i18n'
-import zoneStore from '@store/filterZone'
+import zoneStore from '@store/ws/zone'
 import { Button } from '@ui/button'
 import { Icon } from '@ui/icon'
 import { InputSearch } from '@components/input-search'
