+++ src/components/variant/drawer.tsx
@@ -4,8 +4,8 @@ import { observer } from 'mobx-react-lite'
 
 import { IGridLayout } from '@declarations'
 import { SessionStoreManager } from '@core/storage-management/session-store-manager'
-import variantStore from '@store/variant'
-import columnsStore from '@store/wsColumns'
+import columnsStore from '@store/ws/columns'
+import variantStore from '@store/ws/variant'
 import { VariantBody } from './ui/body'
 import { VariantHeader } from './ui/header'
