+++ src/components/variant/ui/header.tsx
@@ -8,8 +8,8 @@ import { useKeydown } from '@core/hooks/use-keydown'
 import { useVariantIndex } from '@core/hooks/use-variant-index'
 import datasetStore from '@store/dataset'
 import dirinfoStore from '@store/dirinfo'
-import variantStore from '@store/variant'
-import columnsStore from '@store/wsColumns'
+import columnsStore from '@store/ws/columns'
+import variantStore from '@store/ws/variant'
 import { Routes } from '@router/routes.enum'
 import { Button } from '@ui/button'
 import { Icon } from '@ui/icon'
