+++ src/components/variant/ui/igv-button.tsx
@@ -6,7 +6,7 @@ import { observer } from 'mobx-react-lite'
 
 import { t } from '@i18n'
 import dirinfoStore from '@store/dirinfo'
-import variantStore from '@store/variant'
+import variantStore from '@store/ws/variant'
 import { Routes } from '@router/routes.enum'
 import { Button } from '@ui/button'
 import { IAttributeDescriptors } from '@service-providers/dataset-level/dataset-level.interface'
