+++ src/pages/filter/dtree/components/modals/components/modal-save-dataset.tsx
@@ -11,8 +11,8 @@ import dirinfoStore from '@store/dirinfo'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
 import filterPresetsStore from '@store/filter-presets'
-import filterZone from '@store/filterZone'
 import operations from '@store/operations'
+import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
 import { Button } from '@ui/button'
 import { Attention } from '@ui/icons/attention'
@@ -85,7 +85,7 @@ export const ModalSaveDataset = observer(() => {
       return
     }
 
-    filterZone.resetAllSelectedItems()
+    zoneStore.resetAllSelectedItems()
     isDone && history.push(`${Routes.WS}?ds=${value}`)
   }
 
@@ -121,7 +121,7 @@ export const ModalSaveDataset = observer(() => {
     datasetStore.clearZone()
     filterStore.reset()
     dtreeStore.resetData()
-    filterZone.resetAllSelectedItems()
+    zoneStore.resetAllSelectedItems()
   }
 
   const handleChange = (name: string) => {
