+++ src/pages/filter/dtree/components/modals/components/modal-view-variants.tsx
@@ -11,7 +11,7 @@ import { t } from '@i18n'
 import datasetStore from '@store/dataset'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
-import variantStore from '@store/variant'
+import variantStore from '@store/ws/variant'
 import { Icon } from '@ui/icon'
 import { Radio } from '@ui/radio'
 import { VariantBody } from '@components/variant/ui/body'
