+++ src/pages/ws/page.tsx
@@ -8,7 +8,7 @@ import { useParams } from '@core/hooks/use-params'
 import datasetStore from '@store/dataset'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
-import variantStore from '@store/variant'
+import variantStore from '@store/ws/variant'
 import { ExportPanelModal } from '@components/export-panel-modal'
 import { ExportReportButton } from '@components/export-report-button'
 import { Header } from '@components/header'
