+++ src/pages/ws/ui/control-panel/customize-table-modal/columns-list/columns-list.store.ts
@@ -1,7 +1,7 @@
 import { makeAutoObservable } from 'mobx'
 
 import { IColumns } from '@declarations'
-import columnsStore, { columnsToIgnore } from '@store/wsColumns'
+import columnsStore, { columnsToIgnore } from '@store/ws/columns'
 
 export class ColumnListStore {
   public get filteredColumns(): IColumns[] {
