+++ src/pages/ws/ui/control-panel/zone-modals/components/zone-modal-list.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react-lite'
 
-import zoneStore from '@store/filterZone'
+import zoneStore from '@store/ws/zone'
 import { Checkbox } from '@ui/checkbox/checkbox'
 import { MainTableDataCy } from '@components/data-testid/main-table.cy'
