+++ src/pages/ws/ui/table/cell-gene.tsx
@@ -7,7 +7,7 @@ import { observer } from 'mobx-react-lite'
 import { ViewTypeEnum } from '@core/enum/view-type-enum'
 import { geneColorMap } from '@core/gene-color-map'
 import datasetStore from '@store/dataset'
-import columnsStore from '@store/wsColumns'
+import columnsStore from '@store/ws/columns'
 import { CellI } from './cell-interfaces'
 import { PlusIcon } from './plus-icon'
 import { RowHeight } from './table'
