+++ src/pages/ws/ui/table/table.tsx
@@ -17,9 +17,9 @@ import { tableColumnMap } from '@core/table-column-map'
 import { t } from '@i18n'
 import datasetStore from '@store/dataset'
 import filterStore from '@store/filter'
-import zoneStore from '@store/filterZone'
-import variantStore from '@store/variant'
-import columnsStore from '@store/wsColumns'
+import columnsStore from '@store/ws/columns'
+import variantStore from '@store/ws/variant'
+import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
 import { Loader } from '@components/loader'
 import { NoResultsFound } from '@components/no-results-found'
